[refactor] Renamed NationListHeader to ItemsListHeader as it's commonly used for all lists.
