+++ __tests__/src/components/common/ItemsListHeader.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import NationListHeader from '../../../../src/components/common/NationListHeader';
+import NationListHeader from '../../../../src/components/common/ItemsListHeader';
 
-test('NationListHeader renders correctly', () => {
+test('ItemsListHeader renders correctly', () => {
   const tree = renderer
     .create(<NationListHeader title='Test header title' />)
     .toJSON();
